+++ editor/src/components/canvas/controls/select-mode/absolute-resize-control.tsx
@@ -34,12 +34,13 @@ import { createInteractionViaMouse } from '../../canvas-strategies/interaction-s
 import type { EdgePosition } from '../../canvas-types'
 import { CSSCursor } from '../../canvas-types'
 import { getAllTargetsUnderAreaAABB, windowToCanvasCoordinates } from '../../dom-lookup'
-import { SmallElementSize, useBoundingBox } from '../bounding-box-hooks'
+import { useBoundingBox } from '../bounding-box-hooks'
 import { CanvasOffsetWrapper } from '../canvas-offset-wrapper'
 import { isZeroSizedElement } from '../outline-utils'
 import { useMaybeHighlightElement } from './select-mode-hooks'
 import { isEdgePositionEqualTo } from '../../canvas-utils'
 import { treatElementAsGroupLike } from '../../canvas-strategies/strategies/group-helpers'
+import { useResizeEdges } from './use-resize-edges'
 
 export const AbsoluteResizeControlTestId = (targets: Array<ElementPath>): string =>
   `${targets.map(EP.toString).sort()}-absolute-resize-control`
@@ -51,10 +52,6 @@ interface AbsoluteResizeControlProps {
 
 export const SizeLabelID = 'SizeLabel'
 export const SizeLabelTestId = 'SizeLabelTestId'
-
-function shouldUseSmallElementResizeControl(size: number, scale: number): boolean {
-  return size <= SmallElementSize / scale
-}
 
 export const AbsoluteResizeControl = controlForStrategyMemoized(
   ({ targets, pathsWereReplaced }: AbsoluteResizeControlProps) => {
@@ -170,126 +167,6 @@ export const AbsoluteResizeControl = controlForStrategyMemoized(
     )
   },
 )
-
-export function useResizeEdges(
-  targets: ElementPath[],
-  params: {
-    onEdgeMouseDown: (position: EdgePosition) => (e: React.MouseEvent<HTMLDivElement>) => void
-    onEdgeMouseMove: (e: React.MouseEvent<HTMLDivElement>) => void
-    onEdgeDoubleClick: (
-      direction: 'horizontal' | 'vertical',
-    ) => (e: React.MouseEvent<HTMLDivElement>) => void
-    cursors?: {
-      top?: CSSCursor
-      left?: CSSCursor
-      bottom?: CSSCursor
-      right?: CSSCursor
-    }
-  },
-) {
-  const scale = useEditorState(
-    Substores.canvasOffset,
-    (store) => store.editor.canvas.scale,
-    'useResizeEdges scale',
-  )
-
-  const topRef = useBoundingBox(targets, (ref, boundingBox) => {
-    const isSmallElement = shouldUseSmallElementResizeControl(boundingBox.height, scale)
-    const lineSize = ResizeMouseAreaSize / scale
-    const height = isSmallElement ? lineSize / 2 : lineSize
-    const offsetLeft = `0px`
-    const offsetTop = `${-lineSize / 2}px`
-
-    ref.current.style.width = boundingBox.width + 'px'
-    ref.current.style.height = height + 'px'
-    ref.current.style.transform = `translate(${offsetLeft}, ${offsetTop})`
-  })
-
-  const leftRef = useBoundingBox(targets, (ref, boundingBox) => {
-    const isSmallElement = shouldUseSmallElementResizeControl(boundingBox.width, scale)
-    const lineSize = ResizeMouseAreaSize / scale
-    const width = isSmallElement ? lineSize / 2 : lineSize
-    const offsetLeft = `${-lineSize / 2}px`
-    const offsetTop = `0px`
-
-    ref.current.style.width = `${width}px`
-    ref.current.style.transform = `translate(${offsetLeft}, ${offsetTop})`
-    ref.current.style.height = boundingBox.height + 'px'
-  })
-
-  const bottomRef = useBoundingBox(targets, (ref, boundingBox) => {
-    const isSmallElement = shouldUseSmallElementResizeControl(boundingBox.height, scale)
-    const lineSize = ResizeMouseAreaSize / scale
-    const height = isSmallElement ? lineSize / 2 : lineSize
-    const offsetLeft = `0px`
-    const offsetTop = isSmallElement ? `0px` : `${-lineSize / 2}px`
-
-    ref.current.style.transform = `translate(${offsetLeft}, ${offsetTop})`
-    ref.current.style.top = boundingBox.height + 'px'
-    ref.current.style.width = boundingBox.width + 'px'
-    ref.current.style.height = height + 'px'
-  })
-
-  const rightRef = useBoundingBox(targets, (ref, boundingBox) => {
-    const isSmallElement = shouldUseSmallElementResizeControl(boundingBox.width, scale)
-    const lineSize = ResizeMouseAreaSize / scale
-    const width = isSmallElement ? lineSize / 2 : lineSize
-    const offsetLeft = isSmallElement ? `0px` : `${-lineSize / 2}px`
-    const offsetTop = `0px`
-
-    ref.current.style.transform = `translate(${offsetLeft}, ${offsetTop})`
-    ref.current.style.left = boundingBox.width + 'px'
-    ref.current.style.width = width + 'px'
-    ref.current.style.height = boundingBox.height + 'px'
-  })
-
-  return {
-    top: (
-      <ResizeEdge
-        ref={topRef}
-        position={{ x: 0.5, y: 0 }}
-        cursor={params.cursors?.top ?? CSSCursor.ResizeNS}
-        direction='horizontal'
-        onMouseDown={params.onEdgeMouseDown({ x: 0.5, y: 0 })}
-        onMouseMove={params.onEdgeMouseMove}
-        onDoubleClick={params.onEdgeDoubleClick('horizontal')}
-      />
-    ),
-    left: (
-      <ResizeEdge
-        ref={leftRef}
-        position={{ x: 0, y: 0.5 }}
-        cursor={params.cursors?.left ?? CSSCursor.ResizeEW}
-        direction='vertical'
-        onMouseDown={params.onEdgeMouseDown({ x: 0, y: 0.5 })}
-        onMouseMove={params.onEdgeMouseMove}
-        onDoubleClick={params.onEdgeDoubleClick('vertical')}
-      />
-    ),
-    bottom: (
-      <ResizeEdge
-        ref={bottomRef}
-        position={{ x: 0.5, y: 1 }}
-        cursor={params.cursors?.bottom ?? CSSCursor.ResizeNS}
-        direction='horizontal'
-        onMouseDown={params.onEdgeMouseDown({ x: 0.5, y: 1 })}
-        onMouseMove={params.onEdgeMouseMove}
-        onDoubleClick={params.onEdgeDoubleClick('horizontal')}
-      />
-    ),
-    right: (
-      <ResizeEdge
-        ref={rightRef}
-        position={{ x: 1, y: 0.5 }}
-        cursor={params.cursors?.right ?? CSSCursor.ResizeEW}
-        direction='vertical'
-        onMouseDown={params.onEdgeMouseDown({ x: 1, y: 0.5 })}
-        onMouseMove={params.onEdgeMouseMove}
-        onDoubleClick={params.onEdgeDoubleClick('vertical')}
-      />
-    ),
-  }
-}
 
 interface ResizePointProps {
   cursor: CSSCursor
@@ -408,36 +285,6 @@ const ResizePoint = React.memo(
   }),
 )
 ResizePoint.displayName = 'ResizePoint'
-
-interface ResizeEdgeProps {
-  cursor: CSSCursor
-  direction: 'horizontal' | 'vertical'
-  position: EdgePosition
-  onMouseDown: (e: React.MouseEvent<HTMLDivElement>) => void
-  onMouseMove: (e: React.MouseEvent<HTMLDivElement>) => void
-  onDoubleClick: (e: React.MouseEvent<HTMLDivElement>) => void
-}
-
-const ResizeMouseAreaSize = 10
-const ResizeEdge = React.memo(
-  React.forwardRef<HTMLDivElement, ResizeEdgeProps>((props, ref) => {
-    return (
-      <div
-        ref={ref}
-        style={{
-          position: 'absolute',
-          backgroundColor: 'transparent',
-          cursor: props.cursor,
-          pointerEvents: 'initial',
-        }}
-        onMouseDown={props.onMouseDown}
-        onMouseMove={props.onMouseMove}
-        onDoubleClick={props.onDoubleClick}
-        data-testid={ResizePointTestId(props.position)}
-      />
-    )
-  }),
-)
 
 const sizeLabel = (state: FixedHugFill['type'], actualSize: number): string => {
   switch (state) {
